requirements.txt
# This is the minimal set of components required and using the latest
#  versions


aws-shell # Only needed for deploying the site
breathe
gitpython
sphinx
sphinx-rtd-theme
